Add MUD Transparency SBOM Extension support
A draft RFC for a (software) transparency extension to MUD is being
worked on by Eliot Lear at https://github.com/elear/mud-sbom. This
PR adds support for the extension to mudyang.go, based on the YANG
module available in that repository (which can still change). We
also include some of the examples from the repository for demo
purposes.